Rewrite aggregation on interned symbols with once-resolved ops
FindAggregate.Function is now an interned query.Symbol. The parser
resolves the function name once to the pre-interned SymCount, SymSum,
SymAvg, SymMin, SymMax (interned beside SymDollar and re-interned by
ClearInterns); every downstream check is pointer equality, and no
string dispatch remains in any tuple loop. The qb builders carry the
symbols directly.

resolveAggregate maps a symbol to aggregateOps -- an update and a
result function, resolved once per query; an unknown symbol is a loud
error instead of a silent nil aggregate. aggregateAccumulator is flat
per-group value state allocated as one slice. The batch paths fold
collected values through foldAggregateValues; the streaming path
updates accumulators in place. Both grouped tails keep the first
error, so an aggregate finalization error is not clobbered by a clean
scan. The five inline copies of result-symbol construction collapse
into aggregateResultSymbols.

Sums preserve integer typing: an all-int64 group sums to an exact
int64, with no float64 round-trip collapsing adjacent values above
2^53; the first float input promotes the accumulator to float64. sum
and avg error when the result is NaN (Inf and -Inf inputs cancel).
The batch/streaming result-type divergence for integer sums is gone.

Delete the dead AggregateFunction family (query/aggregate.go),
parseAggregate, and the string-switch computeAggregate,
computeAggregateValues, and toFloat64 in executor_utils.go.

+++ datalog/storage/ohlc_realistic_test.go
@@ -173,8 +173,7 @@ func TestOHLCRealisticQueries(t *testing.T) {
 		maxHigh := tuple[4].(float64)
 		_ = tuple[5].(float64) // minLow
 		_ = tuple[6].(float64) // maxClose
-		// BUG: sum should return int64 for int64 input, but returns float64
-		sumVolume := int64(tuple[7].(float64))
+		sumVolume := tuple[7].(int64)
 
 		assert.Equal(t, int64(2025), year)
 		assert.Equal(t, int64(8), month)
